+++ components/Icons/index.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import './DmpIcon.scss';
 
 
 interface IconProps {
@@ -7,6 +8,13 @@ interface IconProps {
 
 export function DmpIcon({ icon }: IconProps) {
   return (
-    <span className="material-symbols-outlined">{icon}</span>
+    <svg
+      className={`dmp-icon icon-${icon}`}
+      width="24px"
+      height="24px"
+      fill="#5f6368"
+    >
+      <use href={`/icons/iconset.svg#icon-${icon}`} />
+    </svg>
   )
 }